Updates docstring
